fix($node/s3): #28 - Fix isValidPolicy fail on size check.
It looks like the `content-length-range` header in the policy document is only
sent when the size validators are enabled client-side.

If any of these are not enabled client-side, then `parsedMaxSize` in
`isPolicyValid()` is `undefined` because the uploader is not sending any
`content-length-range`. This negates the boolean expression at the return.

This fix does two things:
- Makes it so the user can comment out the `expectedMaxSize` variable to do
away with policy document size validation.
- Adds the `expectedMinSize` variable to also validate on the minimum size.
Since minimum size is always sent when maximum size is sent then it seems
to make sense to validate both.